+++ blog/src/main/java/com/example/blog/service/user/UserService.java
@@ -9,6 +9,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -20,6 +21,7 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 @Configuration
 @EnableMethodSecurity
+@EnableWebSecurity
 @Service
 public class UserService implements IUserService {
     @Autowired
@@ -48,41 +50,5 @@ public class UserService implements IUserService {
     public Page<User> getUsers(int page, int size) {
         return iUserRepository.findAll(PageRequest.of(page, size));
     }
-
-    @Bean
-    public UserDetailsService userDetailsService(PasswordEncoder encoder) {
-        UserDetails admin = org.springframework.security.core.userdetails.User.withUsername("admin")
-                .password(encoder.encode("123"))
-                .roles("ADMIN")
-                .build();
-        UserDetails user = org.springframework.security.core.userdetails.User.withUsername("user")
-                .password(encoder.encode("123"))
-                .roles("USER")
-                .build();
-        return new InMemoryUserDetailsManager(admin, user);
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
-        return http
-                .authorizeRequests() // Sử dụng authorizeRequests()
-                .requestMatchers("/").permitAll()
-                .requestMatchers("/user/**").authenticated()
-
-
-                .requestMatchers("/create/**").authenticated()
-
-                .and()
-                .formLogin(formLogin -> formLogin
-//                        .loginPage("/login") nhập
-                        .permitAll())
-                .build();
-    }
-
-
 
 }
